# webui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeurAI",
    "description": "Persian-first, offline-capable meeting intelligence server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/meetings": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Meetings",
        "operationId": "list_meetings_api_meetings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeetingOut"
                  },
                  "type": "array",
                  "title": "Response List Meetings Api Meetings Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Create Meeting",
        "operationId": "create_meeting_api_meetings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Get Meeting",
        "operationId": "get_meeting_api_meetings__meeting_id__get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "meetings"
        ],
        "summary": "Delete Meeting",
        "description": "True deletion (D4): removes the transcript, audio, embeddings, and\nsearch-index entries together — not just the DB row.",
        "operationId": "delete_meeting_api_meetings__meeting_id__delete",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/start": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Start Meeting",
        "operationId": "start_meeting_api_meetings__meeting_id__start_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/stop": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Stop Meeting",
        "operationId": "stop_meeting_api_meetings__meeting_id__stop_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/transcript": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Get Transcript",
        "operationId": "get_transcript_api_meetings__meeting_id__transcript_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SegmentOut"
                  },
                  "title": "Response Get Transcript Api Meetings  Meeting Id  Transcript Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/speakers/relabel": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Relabel Speaker",
        "description": "Manual relabel (D2): name propagates to all of that speaker's segments.",
        "operationId": "relabel_speaker_api_meetings__meeting_id__speakers_relabel_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelabelBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/audio": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Get Audio",
        "operationId": "get_audio_api_meetings__meeting_id__audio_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/bookmarks": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Add Bookmark",
        "operationId": "add_bookmark_api_meetings__meeting_id__bookmarks_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Bookmarks",
        "operationId": "list_bookmarks_api_meetings__meeting_id__bookmarks_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/notes": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Add Note",
        "operationId": "add_note_api_meetings__meeting_id__notes_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Notes",
        "operationId": "list_notes_api_meetings__meeting_id__notes_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/summaries": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Summaries",
        "operationId": "list_summaries_api_meetings__meeting_id__summaries_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/exports/{filename}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Download Export",
        "operationId": "download_export_api_meetings__meeting_id__exports__filename__get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/series": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "List Series",
        "operationId": "list_series_api_series_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "series"
        ],
        "summary": "Create Series",
        "operationId": "create_series_api_series_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookmarkCreate": {
        "properties": {
          "t_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "T Ms"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "t_ms"
        ],
        "title": "BookmarkCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MeetingCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "capture_mode": {
            "type": "string",
            "pattern": "^(room|participants)$",
            "title": "Capture Mode",
            "default": "room"
          },
          "language": {
            "type": "string",
            "pattern": "^(fa|en)$",
            "title": "Language",
            "default": "fa"
          },
          "allow_cloud": {
            "type": "boolean",
            "title": "Allow Cloud",
            "default": false
          },
          "series_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "sensitivity": {
            "type": "string",
            "pattern": "^(normal|confidential)$",
            "title": "Sensitivity",
            "default": "normal"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "MeetingCreate"
      },
      "MeetingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "capture_mode": {
            "type": "string",
            "title": "Capture Mode"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "allow_cloud": {
            "type": "boolean",
            "title": "Allow Cloud"
          },
          "sensitivity": {
            "type": "string",
            "title": "Sensitivity"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "capture_mode",
          "status",
          "language",
          "allow_cloud",
          "sensitivity",
          "series_id",
          "started_at",
          "ended_at",
          "created_at"
        ],
        "title": "MeetingOut"
      },
...
      "NoteCreate": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "t_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "T Ms"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "NoteCreate"
      },
      "RelabelBody": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "label",
          "display_name"
        ],
        "title": "RelabelBody"
      },
...
      "SegmentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pass_name": {
            "type": "string",
            "title": "Pass Name"
          },
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          },
          "speaker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pass_name",
          "start_ms",
          "end_ms",
          "speaker",
          "text"
        ],
        "title": "SegmentOut"
      },
      "SeriesCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "SeriesCreate"
      },
...
    }
  }
}